Update cli code location
