+++ src/index.ts
@@ -1,144 +1,70 @@
-#!/usr/bin/env node
-import { stdout } from "process";
-import { getConfig } from "./config/index.js";
-import { genWasmCode } from "./wasm-code-gen.js";
-import { run } from "./run.js";
-import { parseDirectory, parseFile, parseStd, stdPath } from "./lib/index.js";
-import {
-  expandSyntaxMacrosOfFiles,
-  expandSyntaxMacros,
-} from "./syntax-macros/index.js";
+import { ParsedFiles, parseDirectory } from "./lib/parse-directory.js";
+import { parseFile } from "./lib/parse-file.js";
+import { parseStd, stdPath } from "./lib/parse-std.js";
+import { resolveSrc } from "./lib/resolve-src.js";
+import { File } from "./lib/file.js";
+import { parse } from "./parser.js";
+import { expandSyntaxMacrosOfFiles } from "./syntax-macros/index.js";
 import { resolveFileModules } from "./modules.js";
 import { expandRegularMacros } from "./regular-macros.js";
 import { typeCheck } from "./semantics/index.js";
 import binaryen from "binaryen";
-import { resolveSrc } from "./lib/resolve-src.js";
-import { testGc } from "./lib/binaryen-gc/test.js";
+import { genWasmCode } from "./wasm-code-gen.js";
 
-main().catch(errorHandler);
+export type ParsedModule = {
+  files: ParsedFiles;
+  /** Path to src directory (a folder containing index.void that acts as entry) if available */
+  srcPath?: string;
+  /** Path to root void file */
+  indexPath: string;
+};
 
-async function main() {
-  const config = getConfig();
+export const compileText = async (text: string) => {
+  const parsedModule = await parseText(text);
+  return compileParsedModule(parsedModule);
+};
 
-  if (config.emitParserAst) {
-    return emit(await getParserAst(config.index));
-  }
+export const compilePath = async (path: string) => {
+  const parsedModule = await parsePath(path);
+  return compileParsedModule(parsedModule);
+};
 
-  if (config.emitCoreAst) {
-    return emit(await getCoreAst(config.index));
-  }
+export const compileParsedModule = (module: ParsedModule): binaryen.Module => {
+  const syntaxExpandedFiles = expandSyntaxMacrosOfFiles(module.files);
+  const moduleResolvedModule = resolveFileModules({
+    ...module,
+    files: syntaxExpandedFiles,
+    stdPath: stdPath,
+  });
+  const regularMacroExpandedModule = expandRegularMacros(moduleResolvedModule);
+  const typeCheckedModule = typeCheck(regularMacroExpandedModule);
+  return genWasmCode(typeCheckedModule);
+};
 
-  if (config.emitModuleAst) {
-    return emit(await getModuleAst(config.index));
-  }
+export const parseText = async (text: string): Promise<ParsedModule> => {
+  const file = new File(text, "index");
+  return {
+    files: {
+      index: parse(file),
+      ...(await parseStd()),
+    },
+    indexPath: "index",
+  };
+};
 
-  if (config.emitMacroAst) {
-    return emit(await getMacroAst(config.index));
-  }
-
-  if (config.emitWasmText) {
-    return console.log(
-      await getWasmText(config.index, config.runBinaryenOptimizationPass)
-    );
-  }
-
-  if (config.emitWasm) {
-    return emitWasm(config.index, config.runBinaryenOptimizationPass);
-  }
-
-  if (config.run) {
-    return runWasm(config.index, config.runBinaryenOptimizationPass);
-  }
-
-  if (config.internalTest) {
-    return testGc();
-  }
-
-  console.log(
-    "I don't know what to do with the supplied options. Maybe try something else ¯_(ツ)_/¯"
-  );
-}
-
-async function getParserAst(index: string) {
-  return parseFile(index);
-}
-
-async function getCoreAst(index: string) {
-  const parserAst = await getParserAst(index);
-  return expandSyntaxMacros(parserAst);
-}
-
-async function getModuleAst(index: string) {
-  const src = await resolveSrc(index);
+export const parsePath = async (path: string): Promise<ParsedModule> => {
+  const src = await resolveSrc(path);
 
   const srcFiles = src.srcRootPath
     ? await parseDirectory(src.srcRootPath)
     : { [src.indexPath]: await parseFile(src.indexPath) };
 
-  const parsedFiles = {
-    ...srcFiles,
-    ...(await parseStd()),
-  };
-
-  const files = expandSyntaxMacrosOfFiles(parsedFiles);
-
-  return resolveFileModules({
-    files,
+  return {
+    files: {
+      ...srcFiles,
+      ...(await parseStd()),
+    },
     srcPath: src.srcRootPath,
     indexPath: src.indexPath,
-    stdPath: stdPath,
-  });
-}
-
-async function getMacroAst(index: string) {
-  const moduleAst = await getModuleAst(index);
-  return expandRegularMacros(moduleAst);
-}
-
-async function getWasmMod(index: string, optimize = false) {
-  const ast = await getMacroAst(index);
-  const checkedAst = typeCheck(ast);
-  const mod = genWasmCode(checkedAst);
-
-  if (optimize) {
-    binaryen.setShrinkLevel(3);
-    binaryen.setOptimizeLevel(3);
-    mod.optimize();
-  }
-
-  return mod;
-}
-
-async function getWasmText(index: string, optimize = false) {
-  const mod = await getWasmMod(index, optimize);
-  return mod.emitText();
-}
-
-async function emitWasm(index: string, optimize = false) {
-  const mod = await getWasmMod(index, optimize);
-
-  if (!mod.validate()) {
-    throw new Error("Module is invalid");
-  }
-
-  stdout.write(mod.emitBinary());
-}
-
-async function runWasm(index: string, optimize = false) {
-  const mod = await getWasmMod(index, optimize);
-
-  if (!mod.validate()) {
-    throw new Error("Module is invalid");
-  }
-
-  run(mod);
-}
-
-function emit(json: any) {
-  console.log(JSON.stringify(json, undefined, 2));
-}
-
-function errorHandler(error: Error) {
-  console.error(error);
-  process.exit(1);
-}
+  };
+};
