+++ src/semantics/index.ts
@@ -1,7 +1,7 @@
 import { checkTypes } from "./check-types.js";
 import { initPrimitiveTypes } from "./init-primitive-types.js";
 import { initEntities } from "./init-entities.js";
-import { TypeChecker } from "./types";
+import { TypeChecker } from "./types.js";
 
 const typePhases: TypeChecker[] = [
   initPrimitiveTypes,
